--- a/dlk/python/dlk/templates/include/matrix/multiplication.h
+++ b/dlk/python/dlk/templates/include/matrix/multiplication.h
@@ -36,36 +36,78 @@ inline void matrix_multiplication_col3(
   auto A_colm = row_major_to_col_major(A);
 
   const int Arows = A.rows();
-  if (Arows % 8 == 4 && Arows > 8) {
-    for (std::size_t i = 0; i < B.cols(); ++i) {
-      const float32x4_t rhs0 = vdupq_n_f32((float)(*B.data(0, i)));
-      const float32x4_t rhs1 = vdupq_n_f32((float)(*B.data(1, i)));
-      const float32x4_t rhs2 = vdupq_n_f32((float)(*B.data(2, i)));
-
-      float32x4_t lhs0 = vld1q_f32(A_colm.data(0, 0));
-      float32x4_t lhs1 = vld1q_f32(A_colm.data(0, 1));
-      float32x4_t lhs2 = vld1q_f32(A_colm.data(0, 2));
-      float32x4_t r;
-      for (std::size_t j = 0; j + 7 < Arows; j += 8) {
+  if (Arows > 8) {
+    if (Arows % 8 == 4) {
+      for (std::size_t i = 0; i < B.cols(); ++i) {
+	const float32x4_t rhs0 = vdupq_n_f32((float)(*B.data(0, i)));
+	const float32x4_t rhs1 = vdupq_n_f32((float)(*B.data(1, i)));
+	const float32x4_t rhs2 = vdupq_n_f32((float)(*B.data(2, i)));
+	
+	float32x4_t lhs0 = vld1q_f32(A_colm.data(0, 0));
+	float32x4_t lhs1 = vld1q_f32(A_colm.data(0, 1));
+	float32x4_t lhs2 = vld1q_f32(A_colm.data(0, 2));
+	float32x4_t r;
+	std::size_t j;
+	for (j = 0; j + 7 < Arows; j += 8) {
+	  float32x4_t lhs20 = vld1q_f32(A_colm.data(j+4, 0));
+	  float32x4_t lhs21 = vld1q_f32(A_colm.data(j+4, 1));
+	  float32x4_t lhs22 = vld1q_f32(A_colm.data(j+4, 2));
+	  r = vmulq_f32(lhs0, rhs0);
+	  r = vmlaq_f32(r, lhs1, rhs1);
+	  r = vmlaq_f32(r, lhs2, rhs2);
+	  vst1q_f32(C.data(j, i), r);
+	  lhs0 = vld1q_f32(A_colm.data(j+8, 0));
+	  lhs1 = vld1q_f32(A_colm.data(j+8, 1));
+	  lhs2 = vld1q_f32(A_colm.data(j+8, 2));
+	  r = vmulq_f32(lhs20, rhs0);
+	  r = vmlaq_f32(r, lhs21, rhs1);
+	  r = vmlaq_f32(r, lhs22, rhs2);
+	  vst1q_f32(C.data(j+4, i), r);
+	}
+	r = vmulq_f32(lhs0, rhs0);
+	r = vmlaq_f32(r, lhs1, rhs1);
+	r = vmlaq_f32(r, lhs2, rhs2);
+	vst1q_f32(C.data(Arows, i), r);      
+      }
+    } else if (Arows % 8 == 0) {
+      for (std::size_t i = 0; i < B.cols(); ++i) {
+	const float32x4_t rhs0 = vdupq_n_f32((float)(*B.data(0, i)));
+	const float32x4_t rhs1 = vdupq_n_f32((float)(*B.data(1, i)));
+	const float32x4_t rhs2 = vdupq_n_f32((float)(*B.data(2, i)));
+	
+	float32x4_t lhs0 = vld1q_f32(A_colm.data(0, 0));
+	float32x4_t lhs1 = vld1q_f32(A_colm.data(0, 1));
+	float32x4_t lhs2 = vld1q_f32(A_colm.data(0, 2));
+	float32x4_t r;
+	std::size_t j;
+	for (j = 0; j + 7 < Arows - 8; j += 8) {
+	  float32x4_t lhs20 = vld1q_f32(A_colm.data(j+4, 0));
+	  float32x4_t lhs21 = vld1q_f32(A_colm.data(j+4, 1));
+	  float32x4_t lhs22 = vld1q_f32(A_colm.data(j+4, 2));
+	  r = vmulq_f32(lhs0, rhs0);
+	  r = vmlaq_f32(r, lhs1, rhs1);
+	  r = vmlaq_f32(r, lhs2, rhs2);
+	  vst1q_f32(C.data(j, i), r);
+	  lhs0 = vld1q_f32(A_colm.data(j+8, 0));
+	  lhs1 = vld1q_f32(A_colm.data(j+8, 1));
+	  lhs2 = vld1q_f32(A_colm.data(j+8, 2));
+	  r = vmulq_f32(lhs20, rhs0);
+	  r = vmlaq_f32(r, lhs21, rhs1);
+	  r = vmlaq_f32(r, lhs22, rhs2);
+	  vst1q_f32(C.data(j+4, i), r);
+	}
 	float32x4_t lhs20 = vld1q_f32(A_colm.data(j+4, 0));
 	float32x4_t lhs21 = vld1q_f32(A_colm.data(j+4, 1));
 	float32x4_t lhs22 = vld1q_f32(A_colm.data(j+4, 2));
 	r = vmulq_f32(lhs0, rhs0);
 	r = vmlaq_f32(r, lhs1, rhs1);
 	r = vmlaq_f32(r, lhs2, rhs2);
-	vst1q_f32(C.data(j, i), r);
-	lhs0 = vld1q_f32(A_colm.data(j+8, 0));
-	lhs1 = vld1q_f32(A_colm.data(j+8, 1));
-	lhs2 = vld1q_f32(A_colm.data(j+8, 2));
+	vst1q_f32(C.data(Arows-4, i), r);
 	r = vmulq_f32(lhs20, rhs0);
 	r = vmlaq_f32(r, lhs21, rhs1);
 	r = vmlaq_f32(r, lhs22, rhs2);
-	vst1q_f32(C.data(j+4, i), r);
+	vst1q_f32(C.data(Arows, i), r);           	
       }
-      r = vmulq_f32(lhs0, rhs0);
-      r = vmlaq_f32(r, lhs1, rhs1);
-      r = vmlaq_f32(r, lhs2, rhs2);
-      vst1q_f32(C.data(Arows, i), r);      
     }
   } else if (Arows % 4 == 0) {
     for (std::size_t i = 0; i < B.cols(); i++) {