--- a/dlk/python/dlk/templates/include/matrix/multiplication.h
+++ b/dlk/python/dlk/templates/include/matrix/multiplication.h
@@ -108,6 +108,23 @@ inline void matrix_multiplication_col3(
 	r = vmlaq_f32(r, lhs22, rhs2);
 	vst1q_f32(C.data(Arows, i), r);           	
       }
+    } else {
+      for (std::size_t i = 0; i < B.cols(); i++) {
+	float32x4_t rhs0 = vdupq_n_f32((float)(*B.data(0, i)));
+	float32x4_t rhs1 = vdupq_n_f32((float)(*B.data(1, i)));
+	float32x4_t rhs2 = vdupq_n_f32((float)(*B.data(2, i)));
+	for (std::size_t j = 0; j + 3 < Arows; j += 4) {
+	  float32x4_t lhs0 = vld1q_f32(A_colm.data(j, 0));
+	  float32x4_t lhs1 = vld1q_f32(A_colm.data(j, 1));
+	  float32x4_t lhs2 = vld1q_f32(A_colm.data(j, 2));
+	  
+	  float32x4_t r;
+	  r = vmulq_f32(lhs0, rhs0);
+	  r = vmlaq_f32(r, lhs1, rhs1);
+	  r = vmlaq_f32(r, lhs2, rhs2);
+	  vst1q_f32(C.data(j, i), r);
+	}
+      }
     }
   } else if (Arows % 4 == 0) {
     for (std::size_t i = 0; i < B.cols(); i++) {