--- a/dlk/python/dlk/templates/include/matrix/multiplication.h
+++ b/dlk/python/dlk/templates/include/matrix/multiplication.h
@@ -52,14 +52,14 @@ inline void matrix_multiplication_col3(
 	r = vmlaq_f32(r, lhs2, rhs2);
 	vst1q_f32(C.data(j, i), r);
       }
-    }    
+    }
   } else {
     if (Arows % 8 == 0) {
       for (std::size_t i = 0; i < B.cols(); ++i) {
 	const float32x4_t rhs0 = vdupq_n_f32((float)(*B.data(0, i)));
 	const float32x4_t rhs1 = vdupq_n_f32((float)(*B.data(1, i)));
 	const float32x4_t rhs2 = vdupq_n_f32((float)(*B.data(2, i)));
-	
+
 	float32x4_t lhs0 = vld1q_f32(A_colm.data(0, 0));
 	float32x4_t lhs1 = vld1q_f32(A_colm.data(0, 1));
 	float32x4_t lhs2 = vld1q_f32(A_colm.data(0, 2));
@@ -91,14 +91,14 @@ inline void matrix_multiplication_col3(
 	r = vmulq_f32(lhs20, rhs0);
 	r = vmlaq_f32(r, lhs21, rhs1);
 	r = vmlaq_f32(r, lhs22, rhs2);
-	vst1q_f32(C.data(Arows-4, i), r);           	
+	vst1q_f32(C.data(Arows-4, i), r);
       }
     } else if (Arows % 8 == 4) {
       for (std::size_t i = 0; i < B.cols(); ++i) {
 	const float32x4_t rhs0 = vdupq_n_f32((float)(*B.data(0, i)));
 	const float32x4_t rhs1 = vdupq_n_f32((float)(*B.data(1, i)));
 	const float32x4_t rhs2 = vdupq_n_f32((float)(*B.data(2, i)));
-	
+
 	float32x4_t lhs0 = vld1q_f32(A_colm.data(0, 0));
 	float32x4_t lhs1 = vld1q_f32(A_colm.data(0, 1));
 	float32x4_t lhs2 = vld1q_f32(A_colm.data(0, 2));
@@ -123,7 +123,7 @@ inline void matrix_multiplication_col3(
 	r = vmulq_f32(lhs0, rhs0);
 	r = vmlaq_f32(r, lhs1, rhs1);
 	r = vmlaq_f32(r, lhs2, rhs2);
-	vst1q_f32(C.data(Arows, i), r);      
+	vst1q_f32(C.data(Arows-4, i), r);
       }
     } else {
       for (std::size_t i = 0; i < B.cols(); i++) {
@@ -134,7 +134,7 @@ inline void matrix_multiplication_col3(
 	  float32x4_t lhs0 = vld1q_f32(A_colm.data(j, 0));
 	  float32x4_t lhs1 = vld1q_f32(A_colm.data(j, 1));
 	  float32x4_t lhs2 = vld1q_f32(A_colm.data(j, 2));
-	  
+
 	  float32x4_t r;
 	  r = vmulq_f32(lhs0, rhs0);
 	  r = vmlaq_f32(r, lhs1, rhs1);