@lm-konda 2018-12-11
before
```
TotalRunTime,302901,  sum:302.901ms
matrix_multiplication,108804,2073,  sum:110.877ms
```
after
```
TotalRunTime,291861,  sum:291.861ms
matrix_multiplication,96022,2070,  sum:98.092ms
```
configuration
```
network
　LmnetV1Quantize
dataset you used
　Caltech101
training configurattion
　epoch is 2. other is default.
hardware
　DE10-nano
how many times you run the benchmark
　As much as average average value is known (around 10 or more)
```